communities list to new screen
