+++ src/views/communities/CommunityCard.tsx
@@ -187,6 +187,7 @@ const styles = StyleSheet.create({
     },
     card: {
         marginHorizontal: 16,
+        minWidth: 303,
         marginBottom: 22,
         // marginTop: 8,
         padding: 0,
